[Store] Separate DistributedStorageConfig from backend
