+++ mooncake-store/include/storage/distributed/distributed_storage_backend.h
@@ -1,12 +1,12 @@
 #pragma once
 
-#include <chrono>
 #include <cstdint>
 #include <memory>
 #include <mutex>
 #include <string>
 #include <vector>
 
+#include "config/distributed_storage_config.h"
 #include "fs_adapter.h"
 #include "replica.h"
 #include "storage/distributed/object_storage_adapter.h"
@@ -19,26 +19,6 @@ namespace mooncake {
 enum class DistributedStorageMode {
     kFileSystem,
     kObjectStorage,
-};
-
-struct DistributedStorageConfig {
-    std::string fsdir = "/mnt/3fs/mooncake";
-    std::string fs_adapter_type = "hf3fs";
-    bool enable_health_check = false;
-    int shard_count = 64;
-    uint64_t shard_capacity = 4ULL * 1024 * 1024 * 1024;
-    uint64_t alignment = 4096;
-    bool single_tenant = true;
-    bool eviction_enabled = true;
-    double eviction_high_watermark = 0.9;
-    double eviction_low_watermark = 0.7;
-    std::chrono::seconds deferred_free_duration{30};
-    std::chrono::seconds eviction_check_interval{5};
-
-    bool Validate() const;
-    bool ValidateForAllocator() const;
-    static DistributedStorageConfig FromEnvironment();
-    std::string FormatStr() const;
 };
 
 struct DfsWriteRequest {
